Move common exception handler methods to GlobalExceptionHandler
