+++ src/main/java/com/github/jbence1994/webshop/photo/ProductPhotoControllerExceptionHandler.java
@@ -14,19 +14,4 @@ public class ProductPhotoControllerExceptionHandler {
     public ResponseEntity<ErrorDto> handleProductNotFoundException(ProductNotFoundException exception) {
         return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorDto(exception.getMessage()));
     }
-
-    @ExceptionHandler(InvalidFileExtensionException.class)
-    public ResponseEntity<ErrorDto> handleInvalidFileExtensionException(InvalidFileExtensionException exception) {
-        return ResponseEntity.badRequest().body(new ErrorDto(exception.getMessage()));
-    }
-
-    @ExceptionHandler(PhotoUploadException.class)
-    public ResponseEntity<ErrorDto> handlePhotoUploadException(PhotoUploadException exception) {
-        return ResponseEntity.internalServerError().body(new ErrorDto(exception.getMessage()));
-    }
-
-    @ExceptionHandler(PhotoDeletionException.class)
-    public ResponseEntity<ErrorDto> handlePhotoDeletionException(PhotoDeletionException exception) {
-        return ResponseEntity.internalServerError().body(new ErrorDto(exception.getMessage()));
-    }
 }
