+++ src/test/java/com/github/jbence1994/webshop/common/GlobalExceptionHandlerTests.java
@@ -1,5 +1,8 @@
 package com.github.jbence1994.webshop.common;
 
+import com.github.jbence1994.webshop.photo.InvalidFileExtensionException;
+import com.github.jbence1994.webshop.photo.PhotoDeletionException;
+import com.github.jbence1994.webshop.photo.PhotoUploadException;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.ConstraintViolationException;
 import org.junit.jupiter.api.Test;
@@ -22,6 +25,7 @@ import java.util.stream.Stream;
 import static com.github.jbence1994.webshop.common.FieldErrorTestObject.fieldError;
 import static com.github.jbence1994.webshop.common.ObjectErrorTestObject.objectError1;
 import static com.github.jbence1994.webshop.common.ObjectErrorTestObject.objectError2;
+import static com.github.jbence1994.webshop.photo.PhotoTestConstants.JPEG;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.not;
@@ -65,6 +69,33 @@ public class GlobalExceptionHandlerTests {
         assertThat(result.getStatusCode(), equalTo(HttpStatus.BAD_REQUEST));
         assertThat(result.getBody(), not(nullValue()));
         assertThat(result.getBody().error(), equalTo("The file must not be empty."));
+    }
+
+    @Test
+    public void handleInvalidFileExtensionExceptionTest() {
+        var result = globalExceptionHandler.handleInvalidFileExtensionException(new InvalidFileExtensionException(JPEG));
+
+        assertThat(result.getStatusCode(), equalTo(HttpStatus.BAD_REQUEST));
+        assertThat(result.getBody(), not(nullValue()));
+        assertThat(result.getBody().error(), equalTo("Invalid file extension: .jpeg"));
+    }
+
+    @Test
+    public void handlePhotoUploadExceptionTest() {
+        var result = globalExceptionHandler.handlePhotoUploadException(new PhotoUploadException());
+
+        assertThat(result.getStatusCode(), equalTo(HttpStatus.INTERNAL_SERVER_ERROR));
+        assertThat(result.getBody(), not(nullValue()));
+        assertThat(result.getBody().error(), equalTo("The photo could not be uploaded successfully."));
+    }
+
+    @Test
+    public void handlePhotoDeletionExceptionTest() {
+        var result = globalExceptionHandler.handlePhotoDeletionException(new PhotoDeletionException());
+
+        assertThat(result.getStatusCode(), equalTo(HttpStatus.INTERNAL_SERVER_ERROR));
+        assertThat(result.getBody(), not(nullValue()));
+        assertThat(result.getBody().error(), equalTo("The photo could not be deleted successfully."));
     }
 
     @Test
