+++ src/test/java/com/github/jbence1994/webshop/photo/ProductPhotoControllerExceptionHandlerTests.java
@@ -7,7 +7,6 @@ import org.mockito.InjectMocks;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
 
-import static com.github.jbence1994.webshop.photo.PhotoTestConstants.JPEG;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.not;
@@ -27,35 +26,5 @@ public class ProductPhotoControllerExceptionHandlerTests {
         assertThat(result.getStatusCode(), equalTo(HttpStatus.NOT_FOUND));
         assertThat(result.getBody(), not(nullValue()));
         assertThat(result.getBody().error(), equalTo("No product was found with the given ID: #1."));
-    }
-
-    @Test
-    public void handleInvalidFileExtensionExceptionTest() {
-        var result = productPhotoControllerExceptionHandler
-                .handleInvalidFileExtensionException(new InvalidFileExtensionException(JPEG));
-
-        assertThat(result.getStatusCode(), equalTo(HttpStatus.BAD_REQUEST));
-        assertThat(result.getBody(), not(nullValue()));
-        assertThat(result.getBody().error(), equalTo("Invalid file extension: .jpeg"));
-    }
-
-    @Test
-    public void handlePhotoUploadExceptionTest() {
-        var result = productPhotoControllerExceptionHandler
-                .handlePhotoUploadException(new PhotoUploadException());
-
-        assertThat(result.getStatusCode(), equalTo(HttpStatus.INTERNAL_SERVER_ERROR));
-        assertThat(result.getBody(), not(nullValue()));
-        assertThat(result.getBody().error(), equalTo("The photo could not be uploaded successfully."));
-    }
-
-    @Test
-    public void handlePhotoDeletionExceptionTest() {
-        var result = productPhotoControllerExceptionHandler
-                .handlePhotoDeletionException(new PhotoDeletionException());
-
-        assertThat(result.getStatusCode(), equalTo(HttpStatus.INTERNAL_SERVER_ERROR));
-        assertThat(result.getBody(), not(nullValue()));
-        assertThat(result.getBody().error(), equalTo("The photo could not be deleted successfully."));
     }
 }
